requirements.txt
# Core Flask Application
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
python-dotenv==1.0.0

# Data Processing & Analysis
pandas==2.1.1
numpy==1.24.3
scipy==1.11.3

# Machine Learning - Traditional Models
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0


# Data Visualization & Charts
matplotlib==3.7.2
seaborn==0.12.2

# File Handling & Utilities
openpyxl==3.1.2
xlrd==2.0.1

# HTTP Requests (for AI Chat)
requests==2.31.0
urllib3==2.0.4

# Environment Variables
python-decouple==3.8

# Date/Time Utilities
python-dateutil==2.8.2

# JSON Handling
simplejson==3.19.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23



# ===================================
# INSTALLATION NOTES:
# ===================================

# BASIC INSTALLATION (Core functionality):
# pip install Flask pandas numpy scikit-learn xgboost lightgbm requests python-dotenv

# FULL INSTALLATION (With deep learning):
# pip install -r requirements.txt

# MINIMAL INSTALLATION (Without deep learning):
# Comment out tensorflow, torch, torchvision, torchaudio lines above

# GPU SUPPORT (Optional):
# For NVIDIA GPUs, replace tensorflow with:
# tensorflow-gpu==2.13.0
# And install CUDA toolkit separately

# ALTERNATIVE PYTORCH INSTALLATION:
# Visit https://pytorch.org for platform-specific installation commands
# Example for CPU-only: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

# ===================================
# COMPATIBILITY NOTES:
# ===================================

# Python Version: 3.8-3.11 (recommended: 3.10)
# Operating System: Windows, macOS, Linux
# RAM: Minimum 8GB (16GB recommended for deep learning)
# Storage: 2GB free space for dependencies

# ===================================
# TROUBLESHOOTING:
# ===================================

# If you encounter issues:
# 1. Update pip: pip install --upgrade pip
# 2. Install in virtual environment:
#    python -m venv venv
#    source venv/bin/activate  # On Windows: venv\Scripts\activate
#    pip install -r requirements.txt

# 3. For Apple Silicon Macs (M1/M2):
#    Use conda instead of pip for tensorflow:
#    conda install tensorflow

# 4. If deep learning libraries fail:
#    Install core dependencies first:
#    pip install Flask pandas numpy scikit-learn
#    Then add deep learning libraries one by one

# 5. For memory issues during installation:
#    pip install --no-cache-dir -r requirements.txt